+++ _tests_/react-ts-tw/eslint.config.js
@@ -211,6 +211,7 @@ export default [ // eslint-disable-line
             "code-style/simple-call-single-line": "error",
             "code-style/single-argument-on-one-line": "error",
             "code-style/string-property-spacing": "error",
+            "code-style/svg-component-icon-naming": "error",
             "code-style/ternary-condition-multiline": "error",
             "code-style/type-annotation-spacing": "error",
             "code-style/type-format": "error",